add user defined name prefix to enable multiple compilation units
* define GURKE_STEP_NAME_PREFIX (or if this is not defined
  CUKE_OBJECT_PREFIX for cucumber-cpp compatibility) before including
  Step.h
